Unifies JS & JVM implementation of JSR-310
