+++ core/shared/src/main/scala/cron4s/lib/javatime/JavaTemporalInstance.scala
@@ -16,10 +16,10 @@
 
 package cron4s.lib.javatime
 
-import java.time.temporal.{ChronoField, ChronoUnit, Temporal, TemporalField, TemporalUnit}
+import java.time.temporal._
 
-import cron4s.{CronField, CronUnit}
 import cron4s.datetime.IsDateTime
+import cron4s.{CronField, CronUnit}
 
 import scala.util.{Success, Try}
