+++ docs/src/main/tut/docs/builtin_libs.md
@@ -15,5 +15,5 @@ These are the libraries that are currently supported:
 ### ScalaJS
 
  * JavaScript `Date` API: Package `cron4s.lib.js`
- * JSR-310 Backport: Package `cron4s.lib.threetenbp`. _Support is provided via [scala-java-time](https://github.com/cquiroz/scala-java-time) 
- and requires declaring an explicit dependency in client code, which can also be used in the JVM._
+ * Java Time API: Package `cron4s.lib.javatime`. _Support is provided via [scala-java-time](https://github.com/cquiroz/scala-java-time) 
+ and requires declaring an explicit dependency in client code._
